pyproject.toml
[project]
name = "onebowl2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.79.0",
    "beautifulsoup4>=4.14.3",
    "flask>=3.1.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "recipe-scrapers>=15.11.0",
]
